Add support for blueprints defined outside of pagure
This is the basis for bringing in support for 3rd party extensions to
pagure.

Why we used a different configuration file?
The content of the configuration file will be something like:
````
from pagure_taiga import taiga
PLUGINS = [taiga.TAIGA_NS]
````
It will import the plugin to get its blueprint and store it in
PLUGINS.

There are high chances that the plugin we import will also import
pagure's sources (for example to access the forms or simply the
general configuration).

So we had this defined in the main configuration file we would very
quickly end up with circular imports that would be hard to untangle.

Simply using PAGURE_CONFIG and PAGURE_PLUGIN to specify both
configuration files solves that issue.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>